# Optimize the time statistics in the profile's plan time for scan nodes, including total time, average time, maximum and minimum times.