--- a/fe/fe-core/src/main/java/org/apache/doris/common/profile/ProfileSpan.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/common/profile/ProfileSpan.java
@@ -17,8 +17,6 @@
 
 package org.apache.doris.common.profile;
 
-import org.apache.doris.qe.ConnectContext;
-
 public class ProfileSpan implements AutoCloseable {
 
     public interface TimingOperation {