--- a/fe/fe-core/src/main/java/org/apache/doris/common/util/RuntimeProfile.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/common/util/RuntimeProfile.java
@@ -62,6 +62,7 @@ public class RuntimeProfile {
     public static String MIN_TIME_PRE = "min ";
     public static String AVG_TIME_PRE = "avg ";
     public static String SUM_TIME_PRE = "sum ";
+    public static String CNT_TIME_PRE = "count ";
     @SerializedName(value = "counterTotalTime")
     private Counter counterTotalTime;
     @SerializedName(value = "localTimePercent")