… fe/fe-core/src/main/java/org/apache/doris/datasource/SplitAssignment.java
… fe/fe-core/src/main/java/org/apache/doris/datasource/hive/source/HiveScanNode.java
… fe/fe-core/src/main/java/org/apache/doris/qe/ConnectContext.java