### What problem does this PR solve?

Problem Summary: Previously, the plan time in the profile only recorded the time for individual scan nodes, without aggregating the timing statistics across all scan nodes.

### Release note

The purpose of this improvement is to collect timing statistics for all scan nodes within the plan time section of the profile, including the total time, average time, maximum and minimum times across all nodes.

Introduce `NestedStepTimer` to count the time of each step, and calculate the sum, avg, min, max, and count time of each step. Note：Each StepTimer instance must be used by a single thread only. It is not thread-safe for multiple threads to operate on the same StepTimer instance.

Introduce `ProfileSpan` to use the try-with-resource mechanism to encapsulate the calculation statistics time.
e.g.
```
try (ProfileSpan transaction = ProfileSpan.create(summaryProfile, nodeId, "transaction")) {
    sleep(50);
    
    try (ProfileSpan prepare = ProfileSpan.create(summaryProfile, nodeId, "prepare")) {
        sleep(100);
    }
    
    try (ProfileSpan execute = ProfileSpan.create(summaryProfile, nodeId, "execute")) {
        try (ProfileSpan subquery1 = ProfileSpan.create(summaryProfile, nodeId, "subquery1")) {
            sleep(150);
        }
        
        try (ProfileSpan subquery2 = ProfileSpan.create(summaryProfile, nodeId, "subquery2")) {
            sleep(200);
        }
    }
    
    try (ProfileSpan commit = ProfileSpan.create(summaryProfile, nodeId, "commit")) {
        sleep(50);
    }
}
```

### Result
```
-  Plan  Time:  1s288ms
          -  JoinReorder  Time:  N/A
          -  CreateSingleNode  Time:  N/A
          -  QueryDistributed  Time:  N/A
          -  Init  Scan  Node  Time:  sum  99ms,  count  4,  avg  24ms,  max  84ms,  min  2ms
          -  Finalize  Scan  Node  Time:  sum  868ms,  count  4,  avg  217ms,  max  841ms,  min  5ms
              -  Get  Splits  Time:  sum  827ms,  count  4,  avg  206ms,  max  803ms,  min  5ms
                  -  Get  Partitions  Time:  sum  41ms,  count  4,  avg  10ms,  max  41ms,  min  0ms
                  -  Get  Partition  Files  Time:  sum  785ms,  count  4,  avg  196ms,  max  762ms,  min  5ms
              -  Create  Scan  Range  Time:  sum  40ms,  count  4,  avg  10ms,  max  37ms,  min  0ms
```

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [X] Unit Test: `ProfileSpanWithStepTimerTest`.
    - [ ] Manual test (add detailed scripts or steps below)
    - [ ] No need to test or manual test. Explain why:
        - [ ] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [x] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [x] No.
    - [ ] Yes. <!-- Add document PR link here. eg: https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR should merge into -->