@morningman 2024-11-13
run buildall